ref(interfaces): Rename folder DebugImages to ImagesLoaded
