+++ src/sentry/static/sentry/app/components/events/interfaces/imagesLoaded/debugImage.tsx
