+++ src/sentry/static/sentry/app/components/events/interfaces/imagesLoaded/types.tsx
